+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 =========
 
 
+## Version 26 (2025-05-08)
+- make all accessibility coordinate announcements from 1,1 to 8,8 instead of 0,0 to 7,7
+
 ## Version 25 (2025-05-03)
 - Fullscreen toggle option in settings
 - enhance swipe-to-move: continuous swiping